% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSB_functions.R
\name{adaboostTrain}
\alias{adaboostTrain}
\title{Train a set of Adaboost models}
\usage{
adaboostTrain(
  featureMatrix,
  testFold = 5,
  rounds = 10,
  threshold = 0.5,
  iteration = 10,
  mfinal = 100,
  mfinal.sequence = NA
)
}
\arguments{
\item{featureMatrix}{The data.frame that featureMatrix function outputs.}

\item{testFold}{Portion of data.frame to be used as Test set.}

\item{threshold}{The proportion of positive votes to be seen as consensus.}

\item{iteration}{Number of iterations in training round.}

\item{mfinal}{The mfinal value in boosting function. See more in
?adabog::boosting}

\item{mfinal.sequence}{The list of mfinal values used. Length must be the
same as iteration. Only enter when mfinal is not provided.}

\item{round}{Number of training rounds.}
}
\value{
The adaboost model with the highest augmented F1 value and vecters of
all pre-augmented and augmented rounds.
}
\description{
Train Adaboost models for multiple rounds and get model F1 for each round.
}
