% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSB_functions.R
\docType{data}
\name{atom.color}
\alias{atom.color}
\title{Default atom.color}
\format{
An object of class \code{data.frame} with 7 rows and 1 columns.
}
\usage{
atom.color
}
\description{
data.frame for assigning color codes for atoms in show.grid
function. row.names are fixed. col.name = 'value', row.names = atom (H, N, C,
O, S, CA and CB for alpha and beta carbon), value = string for valid colors
(i.e. 'red').
}
\keyword{datasets}
