% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSB_functions.R
\docType{data}
\name{atom.value}
\alias{atom.value}
\title{Default atom.value}
\format{
An object of class \code{data.frame} with 7 rows and 1 columns.
}
\usage{
atom.value
}
\description{
data.frame for assigning numeric values for atoms in form.grid
function. row.names are fixed. col.name = 'value', row.names = atom atom (H, N
, C, O, S, CA and CB for alpha and beta carbon), value = assigned value for
each atom (0~1).
}
\keyword{datasets}
