% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSB_functions.R
\name{augVote}
\alias{augVote}
\title{Get votes from augmented data}
\usage{
augVote(x, no.Aug, threshold = 0.5, positive.label = 1, negative.label = 0)
}
\arguments{
\item{x}{The vector of prediction results.}

\item{no.Aug}{Number of augmented samples, includes the original sample.}

\item{threshold}{The proportion of positive votes to be seen as consensus}

\item{positive.label}{The positive label of the data set (used in
featureMatrix function).}

\item{negative.label}{The negative label of the data set (used in
featureMatrix function).}
}
\value{
A vector of voted results.
}
\description{
This function takes the augmented prediction results and votes for the
consensus prediction result
}
