% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSB_functions_2.R
\name{caretLime}
\alias{caretLime}
\title{Perform feature importance analysis using Lime}
\usage{
caretLime(fit.caret, set, permutation = 100)
}
\arguments{
\item{fit.caret}{Fit model resulting from caretTrain.}

\item{set}{Set of train and test data returned from trainTest.}

\item{permutation}{Number of permutations in LIME}
}
\value{
A LIME explanation for caret
}
\description{
This function returns a feature importance analysis of caret model using LIME.
}
