% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSB_functions_2.R
\name{caretTrain}
\alias{caretTrain}
\title{Call train from caret on a trainTest set}
\usage{
caretTrain(set, method = "adaboost")
}
\arguments{
\item{set}{Set of train and test data returned from trainTest.}

\item{method}{Avaliable methods in caret::train().}
}
\value{
A fit.caret model
}
\description{
This function retruns a caret model of named method using trainTest set data.
}
