% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSB_functions.R
\docType{data}
\name{coord}
\alias{coord}
\title{3D comparison coordinates for 3D-LBP}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 26 rows and 3 columns.
}
\usage{
coord
}
\description{
Functional matrix for 3D-LBP algorithm. Shall not alter.
}
\keyword{datasets}
