% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSB_functions.R
\name{download.pdb}
\alias{download.pdb}
\title{Download vector of PDB files}
\usage{
download.pdb(pdbids, projectFile = "~/", as.file = TRUE)
}
\arguments{
\item{pdbids}{Vector of pdb ids to be downloaded.}

\item{projectFile}{The directory in which the PDB files are downloaded.}

\item{as.file}{If true, creates a /PDBs folder to store downloaded files.}
}
\value{
A vector of the directories of the downloaded PDB files.
}
\description{
This function downloads a vector of PDB files into the given folder, and
returns a vector of the directories of the PDB files.
}
