% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSB_functions.R
\name{featureMatrix}
\alias{featureMatrix}
\title{Form a feature matrix using lists of positive and negative pdb samples}
\usage{
featureMatrix(
  positive,
  negative,
  projectFile = "~/",
  as.file = TRUE,
  balance = FALSE,
  positive.label = 1,
  negative.label = 0,
  rotations = 0,
  size = 32,
  units = 16,
  combine.chains = TRUE,
  show.grid = FALSE,
  n.feature = 68
)
}
\arguments{
\item{positive}{Character vector of positively labeled pdb files.}

\item{negative}{Character vector of negatively labeled pdb files.}

\item{projectFile}{The directory in which the PDB files are downloaded.}

\item{as.file}{If true, creates a /PDBs folder to store downloaded files.}

\item{balance}{If true balances the number of positive and negative samples.}

\item{positive.label}{Label assigned for positive samples.}

\item{negative.label}{Label assigned for negative samples.}

\item{rotations}{Number of rotational augmentation samples added to the
matrix.}

\item{size}{The length of the recognition cube in Angstroms}

\item{units}{The number of blocks in each axis.}

\item{combine.chains}{If true, assign all chains in the pdb file in one grid.
If false, generates a separate grid for each chain.}

\item{show.grid}{If true, outputs an .html file that contains the 3D
representation of the grid. Directory to output 3D representation .html file
is the same as the .pdb file.}

\item{n.feature}{Suggested number of features in the feature matrix. Will be
fixed to multiples of 68 (a result of lbp3D algorithm).}

\item{center.median}{If true, centers each chain (or all chains depending on
combine.chains) to the median of their xyz.}
}
\value{
A data.frame containing the feature matrix of the input samples
}
\description{
This function takes a vector of positively labeled pdb file names and a
negatively labeled pdb file names to generate a feature matrix for machine
learning model formation. pbd files can be randomly rotated to increase
sample size.
}
