% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSB_functions.R
\name{form.grid}
\alias{form.grid}
\title{Form 3D grid for pdb file}
\usage{
form.grid(
  pdb = NA,
  pdb.name = "pdb0",
  pdb.file = NA,
  combine.chains = TRUE,
  size = 32,
  units = 16,
  center.median = TRUE,
  center = NA,
  show.grid = TRUE
)
}
\arguments{
\item{pdb}{pdb structure from bio3d::read.pdb. (if entered, should not enter
pdb.file).}

\item{pdb.name}{Name of the pdb structure entered. Only enter when pdb
argument is used, and pdb.file is not.}

\item{pdb.file}{Directory to the pdb file. (if entered, should not enter pdb).}

\item{combine.chains}{If true, assign all chains in the pdb file in one grid.
If false, generates a separate grid for each chain.}

\item{size}{The length of the recognition cube in Angstroms}

\item{units}{The number of blocks in each axis.}

\item{center.median}{If true, centers each chain (or all chains depending on
combine.chains) to the median of their xyz.}

\item{center}{Designated center for all chains in pdb. Only enter value when
center.median = FALSE.}

\item{show.grid}{If true, outputs an .html file that contains the 3D
representation of the grid. Directory to output 3D representation .html file
is the same as the .pdb file.}
}
\value{
A list of 3d arrays containing the 3D grids of each chain (or all
chains as a whole when combine.chains = TRUE) of the pdb file with attributes
of input parameters.
}
\description{
This function generates a 3D grid for a given pdb structure or pdb file
directory, with each atom shown as the designated atom.value.
}
