% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSB_functions.R
\name{grid3D}
\alias{grid3D}
\title{3D visualization for pdb grid}
\usage{
grid3D(grid, show.cube = TRUE, show.color = TRUE)
}
\arguments{
\item{grid}{Out put of form.grid function. A 3D grid of pdb information.}

\item{show.color}{If true, uses different colors to represent different atoms.
Uses the data set atom.color as index. More information in ?atom.color.}

\item{show.cude}{If true, shows the enclosing cube of recognition area.}
}
\value{
A plotly 3D scatter plot to show the information in a 3D grid.
}
\description{
This function generates a 3D visualization of a 3D grid generated from a
form.grid function.
}
