% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSB_functions.R
\name{lbp2Feature}
\alias{lbp2Feature}
\title{Form a row in feature matrix from lbp}
\usage{
lbp2Feature(lbp, n.feature = 68, label = 1, name = attr(lbp, "name"))
}
\arguments{
\item{lbp}{lbp matrix generated by lbp3D function}

\item{n.feature}{Suggested number of features in the feature matrix. Will be
fixed to multiples of 68 (a result of lbp3D algorithm).}

\item{label}{Label of the sample features. Should be entered as 0 or 1 for
non-factorized models.}

\item{name}{Name of the feature line. Will be used as rowname in feature
matrix. Default is file name.}
}
\value{
A line of lbp features of the given lbp matrix, with the label as the
last element. Also shows the lbp histogram.
}
\description{
This function transforms a single lbp matrix into 1 line in the feature
matrix
}
