% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSB_functions.R
\name{lbp3d}
\alias{lbp3d}
\title{Form LBP matrix from 3D grid}
\usage{
lbp3d(grid)
}
\arguments{
\item{grid}{Out put of form.grid function. A 3D grid of pdb information.}
}
\value{
LBP matrix 2^26 possible levels
}
\description{
Generate the LBP matrix from a 3D grid generated from from grid.
}
