% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSB_functions_2.R
\name{lbpLIME}
\alias{lbpLIME}
\title{Get feature importance with LIME}
\usage{
lbpLIME(set, method = "adaboost", permutation = 100)
}
\arguments{
\item{set}{Set of train and test data returned from trainTest.}

\item{method}{Avaliable methods in caret::train().}

\item{permutation}{Number of permutations in LIME}
}
\value{
A LIME explanation for caret
}
\description{
This function returns a feature importance analysis of caret model using LIME.
}
