% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSB_functions_2.R
\name{plotLIME}
\alias{plotLIME}
\title{Plot LIME feature importance}
\usage{
plotLIME(lbpLIME)
}
\arguments{
\item{lbpLIME}{Output of pbpLIME}
}
\value{
A feature importance plot
}
\description{
This function plots the feature importance from LIME
}
