% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSB_functions.R
\name{rotatePDB}
\alias{rotatePDB}
\title{Rotate a pdb structure}
\usage{
rotatePDB(pdb, degree = c(0, 0, 0))
}
\arguments{
\item{pdb}{pdb structure output from bio3d::read.pdb}

\item{degree}{A vector of 3 numeric elements indicating the rotation degree
along x, y, and z axis.}
}
\value{
A pdb structure with $Atom$x(y,z) values changed
}
\description{
This function takes a pdb structure read by bio3d::read.pdb function and
rotates it.
}
