% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSB_functions.R
\name{trainTest}
\alias{trainTest}
\title{Separate featureMatrix into Train and Test sets}
\usage{
trainTest(
  featureMatrix,
  positive.label = 1,
  negative.label = 0,
  testFold = 5,
  labelAsFactor = TRUE
)
}
\arguments{
\item{featureMatrix}{The data.frame that featureMatrix function outputs.}

\item{positive.label}{The positive label used in featureMatrix function.}

\item{negative.label}{The negative label used in featureMatrix function.}

\item{testFold}{Portion of data.frame to be used as Test set.}

\item{labelAsFactor}{Set the column of Label into factors.}
}
\value{
A list containing the Train and Test sets. no.Aug attribute contains
the number of base and augmented sample in each group.
}
\description{
This function takes the data.frame generated from featureMatrix and separates
the data.frame in two Train and Test sets for cross-validation.
}
